Revert "OCLOMRS-29:Restructure app to improve code maintainability (#19)"
This reverts commit e77aa4d1a256a43099bf5cc081642a7db346577a.
