+++ src/Reducers/Index.js
@@ -1,7 +1,12 @@
-import { applyMiddleware, createStore } from 'redux';
+import { combineReducers, createStore, applyMiddleware } from 'redux';
 import reduxThunk from 'redux-thunk';
 import logger from 'redux-logger';
-import rootReducer from './index';
+import AuthReducers from './AuthReducers';
+
+// combined reducer to give a global state
+const rootReducer = combineReducers({
+  users: AuthReducers,
+});
 
 const saveState = (state) => {
   try {
