+++ src/__tests__/AuthActions.test.js
@@ -1,5 +1,4 @@
-import { AUTHENTICATED, LOGGED_OUT } from '../redux/actions/types';
-import { login, logout } from '../redux/actions/auth/authActionCreators';
+import { login, AUTHENTICATED, logout, LOGGED_OUT } from '../ActionCreators/AuthActionCreators';
 
 describe('Login', () => {
   const response = {
@@ -28,3 +27,4 @@ describe('Logout', () => {
     expect(logout(response)).toEqual(responseData);
   });
 });
+
